refactor: Moved URL path resolution into helper method
